Origin network bug fixed
