+++ dapp/package.json
@@ -29,12 +29,14 @@
   },
   "dependencies": {
     "3box": "^1.9.0",
+    "@apollo/react-hooks": "^3.1.3",
     "@fuse/entities-contracts": "../@fuse/entities-abi",
     "@fuse/roles": "../@fuse/roles",
     "@fuse/token-factory-contracts": "../@fuse/token-factory-abi",
     "@material-ui/core": "^4.2.1",
     "@portis/web3": "^2.0.0-beta.54",
     "@sentry/browser": "^5.7.1",
+    "apollo-boost": "^0.4.4",
     "babel-core": "^6.26.3",
     "babel-eslint": "^8.2.3",
     "babel-loader": "^7.1.4",
@@ -51,6 +53,7 @@
     "file-loader": "^1.1.11",
     "formik": "^1.5.1",
     "foundation-sites": "^6.5.3",
+    "graphql": "^14.5.8",
     "history": "^4.7.2",
     "html-loader": "^0.5.5",
     "html-webpack-plugin": "^3.2.0",
