+++ dapp/src/containers/Root.jsx
@@ -1,43 +1,28 @@
 import React from 'react'
-import PropTypes from 'prop-types'
-import { Provider } from 'react-redux'
+// import PropTypes from 'prop-types'
+// import { Provider } from 'react-redux'
 import { Switch, Route } from 'react-router'
-import { ConnectedRouter } from 'connected-react-router'
 import Oven from 'components/oven/Oven'
 import Wizard from 'components/wizard'
-import DashboardLayout from 'components/dashboard/containers/MainDashboard'
+import DashboardLayout from 'components/dashboard/containers'
 import Web3 from 'containers/Web3'
-import ScrollToTopController from 'containers/ScrollToTopController'
 import Layout from 'components/common/Layout'
 import HomePage from 'components/home/pages/HomePage'
 import Footer from 'components/common/Footer'
-import ErrorBoundary from 'components/common/ErrorBoundary'
 
-const Root = ({ store, history }) => {
+const Root = ({ history }) => {
   return (
-    <Provider store={store}>
-      <ConnectedRouter history={history}>
-        <ScrollToTopController>
-          <ErrorBoundary>
-            <Layout>
-              <Web3 isMobile={history.location.search.includes('isMobile')} />
-              <Switch>
-                <Route exact path='/' component={props => <HomePage {...props} />} />
-                <Route path='/view/issuance/:templateId?' component={props => <Wizard {...props} />} />
-                <Route path='/view/communities' component={props => <Oven {...props} />} />
-                <Route path='/view/community/:address' component={props => <DashboardLayout {...props} />} />
-              </Switch>
-              <Footer />
-            </Layout>
-          </ErrorBoundary>
-        </ScrollToTopController>
-      </ConnectedRouter>
-    </Provider>
+    <Layout>
+      <Web3 isMobile={history.location.search.includes('isMobile')} />
+      <Switch>
+        <Route exact path='/' component={props => <HomePage {...props} />} />
+        <Route path='/view/issuance/:templateId?' component={props => <Wizard {...props} />} />
+        <Route path='/view/communities' component={props => <Oven {...props} />} />
+        <Route path='/view/community/:address' component={props => <DashboardLayout {...props} />} />
+      </Switch>
+      <Footer />
+    </Layout>
   )
-}
-
-Root.propTypes = {
-  store: PropTypes.object.isRequired
 }
 
 export default Root
